Add runner to run pending background migrations
This class finds a pending background migration and runs it. It checks
that the file exists and it's registered as a pending migration, then
runs it.

We have to have a way of indicating to the migration that it should
_actually_ run, so we use a `Runner.running` class variable to track
that. This should be fine as it should only be run as a single-threaded
task, and we're ensuring it's set to false after the run method is
called.