Run specs against multiple Rails versions
We want to ensure that this gem works against newer versions of Rails.
This commit changes the CI workflow to run against Rails 7.0, 7.1, 7.2
and 8.0.

This is done by managing different Gemfiles for each version, stored in
the gemfiles directory.

There were a few changes to the test setup to get other versions
running. Rails 7.1 changed the way we interact with `SchemaMigration`,
in that it's now a method on the ActiveRecord connection. Rails 7.2
changed the way we deal with connections by requiring us to use a
connection pool. These differences are encapsulated in the spec_helper,
and future changes should hopefully be manageable by adapting the few
conditionals.